+++ vortex-array/src/array/sparse/canonical.rs
